pyproject.toml
[tool.poetry]
name = "trading-unview"
version = "0.1.0"
description = "Trading bot for Bybit - spike detection strategy"
authors = ["heydevaslan"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
pybit = "^5.6.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
